[extensions/ui.pl] Save the prefix string before searching. Include the current input line in an isearch.
[TLily/UI/Curses/Input.pm]
Fixed a bug where the entered line was being appended to history
even when it was identical to the previous line.  The code was clearly
trying to avoid this but had an off-by-one error.

search_forward with $args{next_match} was not taking the length of the
search string into account, so it masked off too many characters.

Only save the history entry in save_history_excursion if it has not already
been saved; the way it was, it was still possible to overwrite a history
entry if the right sequence of steps happened.