Added a new curses-like abstraction layer between the curses UI and curses itself.   This layer is called "FoiledAgain", and the new version of the curses UI that uses it is called "TextWindow".
Once it works properly, we can remove the existing Curses UI and make
TextWindow the default.

At the moment, it works well enough with curses to use, but colors and
styles are broken.   Should be easy enough to fix.

In addition, i've added a second FoiledAgain backend for Win32::Console on
windows.   This was the whole point of this project :)

It kind of works now, but there's are a few known problems:

   - scrolling ($CONSOLE->Scroll()) appears to not be working right.  This
     affects delch/insch as well as normal scrolling.  So you see only the bottom
     few lines of the screen right now, unless you do a pgup/pgdn.
   - keyboard input isn't quite right yet.   I think control characters
     might not be working yet.  Haven't looked too carefully yet.
   - No color or styles yet.   Should be fairly straightforward to add.
   - Haven't done any performance tuning at this point.  I suspect that adding
     an offscreen virtual screen buffer would reduce flicker issues, if
     there are any.
   - Still using the win32 polling event loop.   This may or may not be
     necessary.