Fix the bug where resizing the terminal window would disconnect your lily session.
The problem was that $EAGAIN was not getting properly set under
MSWin32 or cygwin.  A window resize would abort the sysread in
TLily::Server::reader, and the function would not recover because
$EAGAIN was incorrect.

Also added emacs temp files to .gitignore, updated the copyright
dates in tlily.PL and tools/build_tlily, and made build_tlily's
command-line parsing more robust and simpler to use.